Prevent admin from deleting themselves or the last remaining admin

+++ src/modules/user/controller/deleteMe.js
@@ -1,4 +1,4 @@
-const { NotFoundError } = require('../../../utils/errors');
+const { NotFoundError, ForbiddenError } = require('../../../utils/errors');
 const { client } = require('../../../config/redisConfig');
 const { deleteCacheKeys } = require('../../../utils/cacheUtils');
 const User = require('../../../../DB/models/userModel');
@@ -8,10 +8,15 @@ const User = require('../../../../DB/models/userModel');
  * @access    Private/admin - user
  */
 const deleteMe = async (userId) => {
-    const user = await User.findByIdAndDelete(userId);
+    const user = await User.findById(userId);
     if(!user) {
         throw new NotFoundError('User not found');
     }
+    if(user.role === 'admin') {
+        throw new ForbiddenError('Admins cannot delete their own accounts')
+    }
+    await User.findByIdAndDelete(userId);
+
     const cacheKey = `user:${userId}`;
     try {
         await client.del(cacheKey);
@@ -19,6 +24,7 @@ const deleteMe = async (userId) => {
     } catch (err) {
         console.error('Cache Deletion Error:', err);
     }
+    
     return 'Your account has been deleted';;
 };
